Move HubSpot helpers to dedicated module
